Add support for receiving logs over the serial port
We drop them at the moment, as I need to figure out how I want to deal
with internal data buses going forward, but it works.

One option going forward is to just add a second bus for log messages,
and pipe that through. Not a huge fan, but it works.

The other option is to extend the "log or telemetry" enum all the way
over the existing bus, and use that to pipe the data over to the GUI
and the data recorder. In theory this guarantees that we preserve
ordering between telemetry and logs, which might be useful.

A further plan I have is to include host-internal messages in the log
stream, using a separate system log level. That would include for
example when a command has been sent, so that we can commit that to a
hard log alongside the remaining data for analysis.

A third option would be to opt into the tracing framework, and send
logs through that. The downside here would be that we'd need our own
subscriber to pull the data back out, and there's no real benefit to
using tracing otherwise.